Update scripts.js
If applied, this commit will have added a new variable called "pokemonList" and assigned it to a blank array. Then three keys will be added to each Pokemon object for the value of name (this will be a string), height (this will be a number), and types (this will be an array of strings).