Update scripts.js
If applied, this commit will have added a for loop to iterate over each object (item) within the "pokemonList" array. The use of "document.write()" within the loop's code will write the Pokemon name on the websites DOM. A string was added to write the Pokemon's height next to its name inside the loop's code.