Update scripts.js
If applied, this commit will have changed the array name of Pokemon in the "repository" variable back to the "pokemonList" variable. All places within the for loop's code that contained the "repository" variable will have changed back to the "pokemonList" variable as well. This change is to make the assignment easier to complete while following the instructions since the array is still referred to "pokemonList" within the task submission exercises.