Update scripts.js
If applied, this commit will have changed the "scripts.js" file by changing the button function to render the inner text of each pokemon name by using "pokemon.name" to replace "placeholder" as the text content of the element. This will result in the "pokemon" argument parameters passing the "pokemon" from the repository to use "pokemon.name" to access the name of each pokemon in the repository.