Update scripts.js
If applied, this commit will have changed the "scripts.js" file by adding a loadList() function as a return key that uses fetch to GET the complete list of Pokémon from "https://pokeapi.co/api/v2/pokemon/". This is done by adding the loadList() function which fetches the "apiUrl" and passing it in the promise function and the result will be the response "json". A forEach loop is then ran on the repsonse "json" which will take the "json.result" which is the entire result of the "apiUrl" which is the main array/main object. For each of these items, a pokemon variable will be mapped into an object which will have two keys with two values. The first key is a name which will equal the "item.name". The second key is "detailsUrl" which will equal the "item.url". The details are the pokemon Url. Once the data is loaded the pokemon will be added from the add function that was previously created.