Update scripts.js
If applied, this commit will have changed the "scripts.js" file by adding the loadDetails() function which has a parameter of "item". The variable "url" is created and is equal to details which is the "item.detailsUrl". When the promise fetches the details, the "url" will be passed in the promise function and the result will be the repsponse "json" which will return "response.jason" and give the function a parameter of "details" and create three variables. For the "item.url" it will equal "details.sprites.front_default", the "item.height" will equal "details.height", and the "item.types" will equal "details.type".