Update scripts.js
If applied, the commit will have changes the "scripts.js" file by changing the showDetails() function to execute the loadDetails() function with the parameter "pokemon" to pass it as an argument. Then it takes the function as a promise and passes the result from the "loadDetails(pokemon)" function and returns the "console.log" with a pareameter of "pokemon".