Update scripts.js
If applied, this commit will have changed the "scripts.js" by using the showDetails() function with the parameter "item" to pass it as an argument. Then it takes the function as a promise and passes the result from the "pokemonRepository.loadDetails()" function with a parameter of "item" then takes the function as a promise to pass the result of the "pokemonRepository.loadDetails(item)" function and returns the console log with the parameter "item". This will allow the details to be shown in the console for each specific item, in this case, is each specific pokemon.