Update scripts.js
If applied, this commit will have changed the "scripts.js" file by changing the "hideModal" function so that the dialog always rejects the promise when the modal is closed. The promise in the "showDialog" function is also adjusted to save the rejected function of the promise in the "dialogPromiseReject" variable so that when the "hideModal" function is called and the variable is set, the promise will be rejected.